MT-22401: move Pagination to shared api types
Decisions:
- Nothing in the payload is campaign-specific; the next paginated resource
  should reuse it rather than redeclare it
